requirements.txt
Django==5.2.8
python-dotenv==1.0.1
Pillow==10.4.0
# requests==2.32.3
gunicorn
psycopg2-binary
dj-database-url
whitenoise
